requirements.txt
aioconsole
bleak
PyYAML
tabulate
tqdm
